requirements.txt
pytest==7.3.1
selenium==4.9.1
requests==2.31.0
charset-normalizer==3.1.0
urllib3==2.0.2
allure-pytest==2.13.2
allure-python-commons==2.13.2




